Crate solana_sbpf

Virtual machine for SBPF programs.

Modules

aligned_memory

Aligned memory

assembler

This module translates eBPF assembly language to binary.

disassembler

Functions in this module are used to handle eBPF programs with a higher level representation, for example to disassemble the code into a human-readable format.

ebpf

This module contains all the definitions related to eBPF, and some functions permitting to manipulate eBPF instructions.

elf

This module relocates a BPF ELF

elf_parser

Dependency-less 64 bit ELF parser

error

This module contains error and result types

insn_builder

Module provides API to create eBPF programs by Rust programming language

interpreter

Interpreter for eBPF programs.

memory_region

This module defines memory regions

program

Common interface for built-in and user supplied programs

static_analysis

Static Byte Code Analysis

verifier

Verifies that the bytecode is valid for the given config.

vm

Virtual machine for eBPF programs.

Macros

declare_builtin_function

Generates an adapter for a BuiltinFunction between the Rust and the VM interface